refactor: correct namings and imports
